requirements.txt
aws-cdk-lib==2.81.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-glue-alpha==2.69.0a0
# aws-cdk.aws_lambda_python_alpha==2.69.0a0
